Game State & Replays (#6)
* Initial grid setup / game state

* Default Pacman agent setup

* PacmanState from PacmanAgentSetup

* Add constants and state_swap_times

* Fix default agent setup

* Basic game state modifiers

* Ghost behavior, but bad code

* In progress

* Finish code for state

* Documentation and warnings

* Use ComputedGrid from AgentSetup

* kind of working

* Clean up

* Many changes

* Improve grid swapping alongside physics/game state

* Add restart functionality

* Fix bugs where Pacbot's logical position was reset on death, and where Pacbot's rounded position doesn't match a walkable position

* Replay soon??

* Fix replays folder for new clones

* Move colors to separate file

* Replay manager is just barely functional

* Replace time crate with chrono

* Fix clippy warnings

* Center the playback buttons

* Add initial WebSocket connection code

* Fix low framerate when (!recording && paused)

The issue was that the pacman_render.write() lock was unnecessarily
being held through the sleep(0.1 sec) call.

* Remove temporary comments

* Replay refactor finished

* Transparent replay pacman

* GUI improvements

* Refactored again, standalone Replay, removed a lot of unwraps

* Refactored again again, support negative playback speed, fix bugs

* Improvements for GUI with replays, incl. save/load, reset on grid change, optional physics recording

* Fix clippy error

---------

+++ src/gui/transforms.rs
@@ -1,6 +1,6 @@
 //! Transforms between coordinate systems (such as grid/logical <=> screen pixels).
 
-use egui::Pos2;
+use eframe::egui::Pos2;
 
 use crate::grid::Wall;
 
@@ -92,7 +92,7 @@ impl Transform {
     ///
     /// ```
     /// use rapier2d::na::Point2;
-    /// use egui::Pos2;
+    /// use eframe::egui::Pos2;
     /// use mdrc_pacbot_util::grid::Wall;
     /// use mdrc_pacbot_util::gui::transforms::Transform;
     ///
@@ -110,8 +110,8 @@ impl Transform {
     /// assert_eq!(left_bottom, Pos2::new(20.0, 310.0));
     /// ```
     pub fn map_wall(&self, wall: &Wall) -> (Pos2, Pos2) {
-        let left_bottom = Pos2::new(wall.left_bottom.x as f32, wall.left_bottom.y as f32);
-        let right_top = Pos2::new(wall.right_top.x as f32, wall.right_top.y as f32);
+        let left_bottom = Pos2::new(wall.left_bottom.x, wall.left_bottom.y);
+        let right_top = Pos2::new(wall.right_top.x, wall.right_top.y);
         (self.map_point(left_bottom), self.map_point(right_top))
     }
 }
